requirements.txt
Flask==1.1.4
pytest==7.0.1
snowflake-connector-python==2.4.6
pytest-mock==3.6.1